requirements.txt
Flask==2.3.2
easywebdav==1.2.0
requests==2.31.0
urllib3==1.26.16
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.1.0
idna==3.4
gunicorn==20.1.0
Flask-APScheduler==1.12.4